requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.9.0
openai>=1.0.0
anthropic>=0.7.0
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0
requests>=2.31.0
python-dotenv>=1.0.0
setuptools>=68.0.0
wheel>=0.41.0 